refactoring authenticate components
